requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
python-jose
requests==2.31.0
pydantic==1.10.11
httpx